backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
python-dotenv
pydantic

azure-ai-textanalytics
openai
